Replace newlines in tweet text with symbol
Sadly, we cannot send a multi-line message to IRC. It's a limitation of
the protocol that nobody seems to want to change. (IRCv3 has message
batches, which can sort of patch over this, but for backwards
compatibility any extensions always stay within the original limits.)

So we'll replace newlines with a Unicode symbol that looks like an Enter
key, keeping the message on one line but indicating the author's intent.