Give basic data for profile links
The `get_url()` function is now merely a dispatcher. Depending on
whether there is a status ID or not, it calls the appropriate output
function to handle the event.

Still a fair bit of duplicated code, but it'll be easier to refactor
later, once the profile feature is fleshed out more.