twitter: refactor output again; make showing quoted tweet configurable
My decision to go in this direction took into account multiple factors.

First, concatenating two tweets on the same line *necessitated* using
`max_messages` on output, which looks bad with prefixed messages.

Second, the `output_prefix` decorator doesn't work on `url` callables,
because of how they're dispatched. And even if it did, it doesn't handle
messages that get split, so it would still look bad.

Third, it's just common sense that not everyone is going to want quoted
tweets shown. Sometimes people only want the main/parent tweet.

So here we are, with this solution that solves the multi-line output
problem AND gives users a config knob to tweak. As a bonus, the control
flow is much cleaner, too.