WIP update for Twitter API v2 (NOT fully functional)
Twitter's v2 API is still in early access, and not all features needed
for this plugin to work completely are available yet.

This is very much a Work In Progress commit, intended to persist my work
on porting to the new API *so far* in case a catastrophic event destroys
my local working directory. It is not ready for a pull request (though I
will likely open one anyway just so it's clear that someone's working on
the transition from Twitter API v1.1 to v2).