Uncap and update dependency ranges
Users should be able to use later versions of dependencies if they need
to, without having to fight `pip` etc.

If a future release of Sopel breaks things, we'll find out about it no
later than when the first user tries to use this plugin with it and
reports the problem.

Likewise for future releases of `tweety-ns`: newer code will almost
always *fix* problems rather than *cause* them. Issue reports and manual
downgrades are there for the rare cases when things don't go well.

Finally, this starts enforcing a `requires-python` range to match up
with Sopel 8.