Trust the `sign_in()` method
Documentation implies that you have to run `app.connect()` to reuse an
existing saved session, but `app.sign_in()` calls the `connect()` method
internally if the session exists and matches the passed-in `username`.

This behavior should already exist in the plugin's current minimum
`tweety-ns` version of 1.1.4, based on commit dates. Upstream doesn't
maintain tags or other version metadata within the source repository,
though, so this is just me taking an educated guess.